pyproject.toml
# SPDX-FileCopyrightText: 2026 none
#
# SPDX-License-Identifier: CC0-1.0

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kapsule"
version = "0.1.0"
description = "Incus-based container management with KDE integration"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "KDE Community", email = "kde-devel@kde.org"}
]
maintainers = [
    {name = "KDE Community", email = "kde-devel@kde.org"}
]
keywords = [
    "containers",
    "incus",
    "lxc",
    "kde",
    "distrobox",
    "development",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "dbus-fast>=4.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.scripts]
kapsule = "kapsule.main:cli"
kap = "kapsule.main:cli"

[project.urls]
Homepage = "https://invent.kde.org/utilities/kapsule"
Documentation = "https://docs.kde.org/trunk5/en/kapsule/kapsule/"
Repository = "https://invent.kde.org/utilities/kapsule.git"
"Bug Tracker" = "https://bugs.kde.org/enter_bug.cgi?product=kapsule"

[tool.setuptools.packages.find]
where = ["src/cli"]
include = ["*"]

[tool.setuptools.package-dir]
kapsule = "src/cli"

[tool.setuptools.package-data]
kapsule = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
include = ["src/daemon"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
